@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
     <RootNamespace>DevBot9.Protocols.Homie</RootNamespace>
     <AssemblyName>DevBot9.Protocols.Homie</AssemblyName>
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
