@@ -4,99 +4,99 @@ using System.Text;
 using System.Threading;
 using Tevux.Protocols.Mqtt;
 
-namespace TestApp {
-    public class HomieTopicFetcher {
-        private readonly NLog.ILogger _log = NLog.LogManager.GetCurrentClassLogger();
-        private readonly MqttClient _mqttClient = new MqttClient();
-        private readonly Dictionary<string, string> _responses = new Dictionary<string, string>();
-        private ChannelConnectionOptions _channelConnectionOptions;
-        private DateTime _timeOfLastUniqueTopic = DateTime.Now;
+namespace TestApp;
+
+public class HomieTopicFetcher {
+    private readonly NLog.ILogger _log = NLog.LogManager.GetCurrentClassLogger();
+    private readonly MqttClient _mqttClient = new();
+    private readonly Dictionary<string, string> _responses = new();
+    private ChannelConnectionOptions _channelConnectionOptions;
+    private DateTime _timeOfLastUniqueTopic = DateTime.Now;
 
 
-        public void Initialize(ChannelConnectionOptions channelOptions) {
-            _channelConnectionOptions = channelOptions;
-            _mqttClient.Initialize();
+    public void Initialize(ChannelConnectionOptions channelOptions) {
+        _channelConnectionOptions = channelOptions;
+        _mqttClient.Initialize();
 
-            _mqttClient.PublishReceived += HandlePublishReceived;
+        _mqttClient.PublishReceived += HandlePublishReceived;
+    }
+
+    public void FetchTopics(string filter, out string[] topics) {
+        _responses.Clear();
+        _mqttClient.ConnectAndWait(_channelConnectionOptions);
+        while (_mqttClient.IsConnected == false) {
+            Thread.Sleep(100);
         }
 
-        public void FetchTopics(string filter, out string[] topics) {
+        _mqttClient.Subscribe(filter, QosLevel.AtLeastOnce);
+
+        Thread.Sleep(2000);
+        _mqttClient.Unsubscribe(filter);
+        _mqttClient.DisconnectAndWait();
+
+        topics = new string[_responses.Count];
+        var i = 0;
+        foreach (var response in _responses) {
+            topics[i] = response.Key + ":" + response.Value;
+            i++;
+        }
+
+        while (_mqttClient.IsConnected) { Thread.Sleep(100); }
+    }
+
+    public void FetchDevices(string baseTopic, out string[] topics) {
+        var allTheTopics = new List<string>();
+
+        _mqttClient.ConnectAndWait(_channelConnectionOptions);
+        while (_mqttClient.IsConnected == false) {
+            Thread.Sleep(100);
+        }
+
+        _responses.Clear();
+        _mqttClient.SubscribeAndWait($"{baseTopic}/+/$homie", QosLevel.AtLeastOnce);
+        Thread.Sleep(500);
+        _mqttClient.Unsubscribe($"{baseTopic}/+/$homie");
+
+        var deviceString = $"Found { _responses.Count} Homie devices: ";
+        var devices = new List<string>();
+        foreach (var deviceTopic in _responses) {
+            var deviceName = deviceTopic.Key.Split('/')[1];
+            devices.Add(deviceName);
+            deviceString += deviceName + " ";
+        }
+        _log.Info(deviceString);
+
+        foreach (var device in devices) {
             _responses.Clear();
-            _mqttClient.ConnectAndWait(_channelConnectionOptions);
-            while (_mqttClient.IsConnected == false) {
+
+            _mqttClient.SubscribeAndWait($"{baseTopic}/{device}/#", QosLevel.AtLeastOnce);
+            while ((DateTime.Now - _timeOfLastUniqueTopic).TotalMilliseconds < 500) {
                 Thread.Sleep(100);
             }
+            _mqttClient.UnsubscribeAndWait($"{baseTopic}/{device}/#");
 
-            _mqttClient.Subscribe(filter, QosLevel.AtLeastOnce);
-
-            Thread.Sleep(2000);
-            _mqttClient.Unsubscribe(filter);
-            _mqttClient.DisconnectAndWait();
-
-            topics = new string[_responses.Count];
-            var i = 0;
-            foreach (var response in _responses) {
-                topics[i] = response.Key + ":" + response.Value;
-                i++;
+            _log.Info($"{_responses.Count} topics for {device}.");
+            foreach (var topic in _responses) {
+                allTheTopics.Add(topic.Key + ":" + topic.Value);
             }
-
-            while (_mqttClient.IsConnected) { Thread.Sleep(100); }
         }
 
-        public void FetchDevices(string baseTopic, out string[] topics) {
-            var allTheTopics = new List<string>();
+        _mqttClient.DisconnectAndWait();
 
-            _mqttClient.ConnectAndWait(_channelConnectionOptions);
-            while (_mqttClient.IsConnected == false) {
-                Thread.Sleep(100);
-            }
+        topics = allTheTopics.ToArray();
 
-            _responses.Clear();
-            _mqttClient.SubscribeAndWait($"{baseTopic}/+/$homie", QosLevel.AtLeastOnce);
-            Thread.Sleep(500);
-            _mqttClient.Unsubscribe($"{baseTopic}/+/$homie");
+        while (_mqttClient.IsConnected) { Thread.Sleep(100); }
+    }
 
-            var deviceString = $"Found { _responses.Count} Homie devices: ";
-            var devices = new List<string>();
-            foreach (var deviceTopic in _responses) {
-                var deviceName = deviceTopic.Key.Split('/')[1];
-                devices.Add(deviceName);
-                deviceString += deviceName + " ";
-            }
-            _log.Info(deviceString);
+    private void HandlePublishReceived(object sender, PublishReceivedEventArgs e) {
+        var payload = Encoding.UTF8.GetString(e.Message);
 
-            foreach (var device in devices) {
-                _responses.Clear();
-
-                _mqttClient.SubscribeAndWait($"{baseTopic}/{device}/#", QosLevel.AtLeastOnce);
-                while ((DateTime.Now - _timeOfLastUniqueTopic).TotalMilliseconds < 500) {
-                    Thread.Sleep(100);
-                }
-                _mqttClient.UnsubscribeAndWait($"{baseTopic}/{device}/#");
-
-                _log.Info($"{_responses.Count} topics for {device}.");
-                foreach (var topic in _responses) {
-                    allTheTopics.Add(topic.Key + ":" + topic.Value);
-                }
-            }
-
-            _mqttClient.DisconnectAndWait();
-
-            topics = allTheTopics.ToArray();
-
-            while (_mqttClient.IsConnected) { Thread.Sleep(100); }
+        if (_responses.ContainsKey(e.Topic) == false) {
+            _responses.Add(e.Topic, payload);
+            _timeOfLastUniqueTopic = DateTime.Now;
         }
-
-        private void HandlePublishReceived(object sender, PublishReceivedEventArgs e) {
-            var payload = Encoding.UTF8.GetString(e.Message);
-
-            if (_responses.ContainsKey(e.Topic) == false) {
-                _responses.Add(e.Topic, payload);
-                _timeOfLastUniqueTopic = DateTime.Now;
-            }
-            else {
-                _responses[e.Topic] = payload;
-            }
+        else {
+            _responses[e.Topic] = payload;
         }
     }
 }
